+++ core/src/Streamly/Data/Unfold.hs
@@ -177,7 +177,6 @@ module Streamly.Data.Unfold
     -- ** From Containers
     , fromList
     , fromListM
-    , fromStream
 
     -- * Combinators
     -- ** Mapping on Input
@@ -212,22 +211,4 @@ import Prelude hiding
     ( concat, map, mapM, takeWhile, take, filter, const, drop, dropWhile
     , zipWith
     )
-import Streamly.Internal.Data.Stream.IsStream.Type (IsStream)
-import Streamly.Internal.Data.Unfold hiding (fromStream)
-
-import qualified Streamly.Internal.Data.Stream.IsStream.Type as IsStream
-import qualified Streamly.Internal.Data.Unfold as Unfold
-
--- XXX Using Unfold.fromStreamD seems to be faster (using cross product test
--- case) than using fromStream even if it is implemented using fromStreamD.
--- Check if StreamK to StreamD rewrite rules are working correctly when
--- implementing fromStream using fromStreamD.
-
--- | Convert a stream into an 'Unfold'. Note that a stream converted to an
--- 'Unfold' may not be as efficient as an 'Unfold' in some situations.
---
--- /Since: 0.8.0/
---
-{-# INLINE_NORMAL fromStream #-}
-fromStream :: (IsStream t, Applicative m) => Unfold m (t m a) a
-fromStream = lmap IsStream.toStream Unfold.fromStreamK
+import Streamly.Internal.Data.Unfold