+++ docs/User/ProjectRelated/Changelog.md
@@ -14,6 +14,9 @@
     writeWithBufferOf -> writeWith
     writeChunksWithBufferOf -> writeChunksWith
 
+* Following API has been moved from Streamly.data.Unfold module:
+    Streamly.data.Unfold.fromStream -> Streamly.Internal.Data.Stream.IsStream.readStream
+
 ### Deprecations
 
 * In `Streamly.Data.Fold`: